refactor: modularize reference collapse
- move reference collapse helpers out of general workflow utilities

- stop persisting collapse stats as a cache manifest
